Restructure classes to editor and runtime-only
We use inheritance for splitting the class for editor and
runtime-only. Editor classes could be also used in runtime, but
they have operations that are used only in edit mode. We don't want
to make solo assembly for editor classes, because we want to use
also the editor classes in runtime for debugging purpose. At least
for now.